+++ Software/emulator/srec.cpp
@@ -9,9 +9,9 @@ extern tgfTextOverlay    con;
 extern char              lineBuffer[255];
 
 
-static int writeMem8( ulong address, ulong data )
+static int writeMem8( uint32_t address, uint32_t data )
 {
-   uchar mask;
+   uint8_t mask;
 
    data = data & 0xff;
 
@@ -87,7 +87,7 @@ static int hexToIByte( char *buf, int position )
    return ( d1 << 4 ) | d2;
 }
 
-static ulong decodeAndDisplayS0Record( char *buf )
+static uint32_t decodeAndDisplayS0Record( char *buf )
 {
    int idx;
    int recordLength;
@@ -112,14 +112,14 @@ static ulong decodeAndDisplayS0Record( char *buf )
    return 0;
 }
 
-static ulong decodeAndSaveS1_2_3Record( char *buf, ulong recordType )
+static uint32_t decodeAndSaveS1_2_3Record( char *buf, uint32_t recordType )
 {
    int   rv;
-   ulong i;
-   ulong dataFieldIdx;
-   ulong address;
-   ulong recordLength;
-   ulong addressFieldLength;
+   uint32_t i;
+   uint32_t dataFieldIdx;
+   uint32_t address;
+   uint32_t recordLength;
+   uint32_t addressFieldLength;
 
    unsigned char *memPtr;
    
@@ -164,7 +164,7 @@ static ulong decodeAndSaveS1_2_3Record( char *buf, ulong recordType )
       if( rv == -1 ) return 1;
 
       address <<= 8;
-      address |= (uchar)rv;
+      address |= (uint8_t)rv;
 
    }
 
@@ -183,12 +183,12 @@ static ulong decodeAndSaveS1_2_3Record( char *buf, ulong recordType )
 }
 
 
-static ulong decodeAndExecuteS7_8_9Record( char *buf, ulong recordType )
+static uint32_t decodeAndExecuteS7_8_9Record( char *buf, uint32_t recordType )
 {
    long  rv;
-   ulong address;
-   ulong i;
-   ulong addressFieldLength;
+   uint32_t address;
+   uint32_t i;
+   uint32_t addressFieldLength;
 
 
    switch( recordType )
@@ -218,18 +218,18 @@ static ulong decodeAndExecuteS7_8_9Record( char *buf, ulong recordType )
       if( rv == -1 ) return 1;
    
       address <<= 8;
-      address |= (uchar)rv;
+      address |= (uint8_t)rv;
    
    }
    
    return address;
 } 
 
-ulong srecLoadFile( char *path )
+uint32_t srecLoadFile( char *path )
 {
    tosFile  in;
-   ulong    rv;
-   ulong    i;
+   uint32_t    rv;
+   uint32_t    i;
 
    if( osFOpen( &in, path, OS_FILE_READ ) )
    {
@@ -239,7 +239,7 @@ ulong srecLoadFile( char *path )
 
    i = 0;
 
-   while( !osFGetS( &in, (uchar*)lineBuffer, sizeof( lineBuffer ) - 1 ) )
+   while( !osFGetS( &in, (uint8_t*)lineBuffer, sizeof( lineBuffer ) - 1 ) )
    {
 
       rv = 0;
